+++ lib/experimental/Information/Avatars/exports.ts
@@ -1,5 +1,4 @@
 export * from "./CompanyAvatar"
 export * from "./PersonAvatar"
-export * from "./render"
 export * from "./TeamAvatar"
-export * from "./types"
+export * from "./utils"
