+++ lib/experimental/Navigation/Dropdown/index.tsx
@@ -2,7 +2,7 @@ import { Button } from "@/components/Actions/Button"
 import {
   AvatarVariant,
   renderAvatar,
-} from "@/experimental/Information/Avatars/exports"
+} from "@/experimental/Information/Avatars/utils"
 import * as Icons from "@/icons/app"
 import { Link } from "@/lib/linkHandler"
 import { cn } from "@/lib/utils"
